+++ src/main/java/com/czertainly/core/auth/oauth2/CzertainlyAuthenticationSuccessHandler.java
@@ -1,0 +1,20 @@
+package com.czertainly.core.auth.oauth2;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Component
+public class CzertainlyAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
+
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
+        String redirectUrl = (String) request.getSession().getAttribute("redirectUrl");
+        request.getSession().removeAttribute("redirectUrl");
+        response.sendRedirect(redirectUrl);
+    }
+}
